rewind stuff

+++ sources/Engine/FizX.Core/Actors/ActorComponents/ActorComponent.cs
@@ -6,7 +6,7 @@ public abstract class ActorComponent : ITickable
 {
     public Actor? Actor { get; internal set; }
 
-    public virtual void Tick(int deltaMs) { }
+    public virtual void Tick(float deltaMs) { }
 
     public override string ToString()
     {
